Huge improvement in clarity, readibility, consistency.
The style is easy for the reader and maintainer.

Look and feel is consistent with very popular, influential and
respected book "Perl Best Practices" by Damian Conway. The book served
as the basis for the later developent of Perl::Critic, the major Perl
static code analyzer.

No need to reinvent the wheel here.