Do not use ivar trim's -f option on SE data
This option filters sequenced PE *fragments* that overlap more than one
amplicon. For SE data, the option just requires extra compute time, but
does not change the output of the tool.

Also removed the mentioning of the amplicon info file from the changelog
since that was not needed in the previous version.