Rewire logic for amplicon bias correction
This changeset makes the amplicon bias correctionmore robust and better
interoperable with the reporting workflow.
The first and second (after amplicon removal) round of variant calling are now
carried out with identical lofreq parameter settings. bcftools annotate
is then used to carry over the bias-corrected call stats to the variant
calls obtained in the first round. At the same time both variant call
lists are filtered with (by default) identical DP and DP_ALT filters as
in the reporting workflow, and stats of filter-passing variants from the first
round that fail to pass after the second round of calling are
transferred back to the first bcftools annotate output.
Together this makes sure that no initially called variant gets lost as a
consequence of amplicon bias correction and that no initially
filter-passing variant gets filtered out after correction. The
AmpliconBias INFO flag is used to mark all such variants for which amplicon
bias correction was not performed.