Filter based on & report unbiased AF / q30 DP_ALT
In this version, the user-supplied AF and DP_ALT criteria are compared
against (DP4[2] + DP4[3]) / DP and AFlofreq * DP, respectively.
The unbiased version of the AF is also what's used in the variant
frequency plot, and what's reported as AF in the combined variant
reports. The "Combined Variant Report by Sample" reports the original
biased AF value as "AFcaller" in addition.