Simplify workflow with new compute tool version
With version 2.0 of the column_maker tool it is possible to chain
column manipulations in a single run of the tool and to insert new
columns in specific places, or to use them to replace existing columns.

In the reporting workflow, this can be used to save two workflow steps.