Add ref genome masking with bedtools back in
This step was dropped as seemingly redundant in release 0.2, but is
important when there are zero consensus variants to be integrated into
the reference since  with an empty input vcf bcftools consensus will silently
not apply any changes to the ref genome, i.e. will also not do any masking.